+++ appconfigservice/api/controllers/config.py
@@ -10,4 +10,4 @@ APP_CONFIG_MAX_POOLSIZE = 100,
 APP_CONFIGS_COLLECTION = 'app_configs'
 API_LOC = os.getenv('API_LOC', '../../')
 # APP_CONFIG_ENDPOINT = os.getenv('APPCONFIG_ENDPOINT', '/rest_service')
-DEBUG = os.getenv('DEBUG', 'False')
+DEBUG = bool(os.getenv('DEBUG', 'False') == 'True')